protected route api

+++ app/api/deleteProduct/route.ts
@@ -2,15 +2,17 @@ import {prisma} from "@/db/db";
 import {NextResponse} from 'next/server'
 import * as fs from "node:fs/promises";
 import {join} from "path";
-
-export async function DELETE(request: Request, {params}: { params: { id: string } }) {
+import {auth} from "@/auth";
+export const DELETE = auth(async function DELETE(request) {
+    if (request.auth?.user?.role !== "ADMIN") return NextResponse.json({message: "Not authenticated"}, {status: 401})
     try {
-        const id = Number(params.id);
+        const formData = await request.formData();
+        const id = Number(formData.get('id'));
         const product = await prisma.products.findUnique({
             where: {id: id}
         })
-        const fullpath = join(process.cwd(),"public",product?.image as string);
-        await fs.unlink(fullpath);
+        const fullPath = join(process.cwd(),"public",product?.image as string);
+        await fs.unlink(fullPath);
         await prisma.products.delete({
             where: {id: id}
         })
@@ -20,4 +22,4 @@ export async function DELETE(request: Request, {params}: { params: { id: string 
         return NextResponse.json({message: error}, {status: 500});
     }
 
-}
+})
